WIP: Add signature verification step for cuemod init. Remove user:pass auth from artifact push
Move create workspace to a dedicated func, and adds signature verification step prompting user input to proceed if verification fails
because of mismatch of signatures. if user decides to continue, creates workspace.
artifact push is not behaving as expected by adding withAuthConfig(username, password) for auth. Hence, removed for timebeing.

Signed-off-by: Santosh Kaluskar <ksantosh@intelops.dev>